refactor: extract duplicated waitForEvent test helper to helpers.ts (#115)
- Create src/test/helpers.ts with waitForEvent and expected* variables
- Update score.test.ts and controls.test.ts to import from helpers
- Fix score.test.ts bug: d.position.choir -> d.position.bar for expectedBar
- Add helpers.test.ts with trivial waitForEvent smoke test
