+++ src/test/controls.test.ts
@@ -1,38 +1,12 @@
 import { MusicControls } from "../ts/MusicControls";
 import config from "../ts/config";
 import { processLilypond, barCount } from "../ts/lily";
-
-var expectedBar: any;
-var expectedChoir: any;
-var expectedPart: any;
-
-// A helper function that allows us to detect events on element
-// of type eventName have been fired
-function waitForEvent(
-  element: HTMLElement,
-  eventName: string,
-  handler: (event: Event) => Promise<any>,
-  c?: any,
-  p?: any,
-  b?: any
-): Promise<any> {
-  expectedChoir = c;
-  expectedPart = p;
-  expectedBar = b;
-  return new Promise<any>((resolve, reject) => {
-    const eventListener = async (event: Event) => {
-      try {
-        const result = await handler(event);
-        resolve(result); // Resolve with handler's result
-      } catch (error) {
-        reject(error); // Reject on error
-      } finally {
-        element.removeEventListener(eventName, eventListener, false);
-      }
-    };
-    element.addEventListener(eventName, eventListener, false);
-  });
-}
+import {
+  expectedBar,
+  expectedChoir,
+  expectedPart,
+  waitForEvent,
+} from "./helpers";
 
 function matchesWildcard(pattern: string, str: string): boolean {
   const regex = new RegExp("^" + pattern.replace(/\*/g, ".*") + "$");
